refactor(test): 优化测试代码中的数据类型和Mock配置
- 将BigInteger类型替换为long基本类型以简化测试数据设置
- 移除不必要的BigInteger导入语句
- 使用lenient()配置Mock行为以避免严格的验证异常
- 统一订单服务测试中的数据类型转换处理

+++ goodskill-order-provider/goodskill-order-service/src/test/java/com/goodskill/order/controller/OrderControllerTest.java
@@ -2,6 +2,7 @@ package com.goodskill.order.controller;
 
 import com.goodskill.core.info.Result;
 import com.goodskill.core.pojo.dto.OrderDTO;
+import com.goodskill.core.util.UserInfoUtil;
 import com.goodskill.order.entity.Order;
 import com.goodskill.order.service.impl.OrderServiceImpl;
 import org.junit.jupiter.api.Test;
@@ -10,17 +11,15 @@ import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.MockedStatic;
 import org.mockito.junit.jupiter.MockitoExtension;
-import com.goodskill.core.util.UserInfoUtil;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageImpl;
 
-import java.math.BigInteger;
 import java.util.Collections;
 import java.util.List;
 import java.util.Map;
 
 import static org.junit.jupiter.api.Assertions.*;
-import static org.mockito.ArgumentMatchers.*;
+import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.*;
 
 /**
@@ -53,7 +52,7 @@ class OrderControllerTest {
     void shouldSaveRecordSuccessfully() {
         // Given
         OrderDTO orderDTO = new OrderDTO();
-        orderDTO.setSeckillId(BigInteger.valueOf(1L));
+        orderDTO.setSeckillId(1L);
         orderDTO.setUserPhone("13800138000");
         when(orderService.saveRecord(any(OrderDTO.class))).thenReturn("order-123");
